Image_Caption_Generator/requirements.txt
numpy~=1.16.4
Keras~=2.2.4
tqdm~=4.28.1
nltk~=3.2.5
matplotlib~=3.0.3
Pillow~=4.3.0
tensorflow~=1.13.1